+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ master
 
 a) [TASK][BREAKING] Update to possible braking symfony/dotenv 5.0 compatibility for "sourcebroker/deployer-extended",
     "sourcebroker/deployer-extended-media", "sourcebroker/deployer-extended-database.
+b) [TASK] Add ddev config.
 
 2.0.2
 ~~~~~