+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ b) [TASK] Add ddev config.
 c) [TASK][BREAKING] Remove auto creation of database and .env file. Use ddev https://ddev.readthedocs.io/en/stable/ or other
    similar projects.
 d) [TASK][BREAKING] Refactor EnvDrive to use Env from sourcebroker/deployer-instance. Use $_ENV instead of getenv() as symfony/dotenv 5.0 compatibility.
+e) [TASK] Add .ddev folder to clear_paths.
 
 2.0.2
 ~~~~~