+++ CHANGELOG.rst
@@ -2,8 +2,8 @@
 Changelog
 ---------
 
-master
-~~~~~~
+3.0.0
+~~~~~
 
 a) [TASK][BREAKING] Update to possible braking symfony/dotenv 5.0 compatibility for "sourcebroker/deployer-extended",
     "sourcebroker/deployer-extended-media", "sourcebroker/deployer-extended-database.